TryParse example.
In this example we have a string that must be converted to an integer. There is a method called tryParseInt. Inside it has a try-catch structure that tries to convert a string to an integer value. If it does, the method returns true, otherwise false. This is important because the program will only run in full if it is possible to extract a number from the string.